pyproject.toml
[project]
name = "carservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-admin-interface>=0.32.0",
    "django-colorfield>=0.14.0",
    "fastapi>=0.135.1",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
